+++ modules/tinymce/src/core/test/ts/browser/fmt/ListItemFormatTest.ts
@@ -1,7 +1,7 @@
 import { Cursors } from '@ephox/agar';
 import { context, describe, it } from '@ephox/bedrock-client';
 import { Type } from '@ephox/katamari';
-import { TinyAssertions, TinyHooks, TinySelections } from '@ephox/wrap-mcagar';
+import { TinyAssertions, TinyHooks, TinySelections, TinyState } from '@ephox/wrap-mcagar';
 
 import Editor from 'tinymce/core/api/Editor';
 import { FormatVars } from 'tinymce/core/fmt/FormatTypes';
@@ -216,16 +216,14 @@ describe('browser.tinymce.core.fmt.ListItemFormatTest', () => {
       );
 
       it('TINY-9563: applying bold on LIs in a noneditable root should not get bold styles', () => {
-        const editor = hook.editor();
-
-        editor.getBody().contentEditable = 'false';
-        testApplyInlineListFormat({
-          format: 'bold',
-          rawInput: '<p>a</p><ul><li>b</li><li>c</li></ul><p>d</p>',
-          selection: { startPath: [ 0, 0 ], soffset: 0, finishPath: [ 2, 0 ], foffset: 1 },
-          expected: '<p>a</p><ul><li>b</li><li>c</li></ul><p>d</p>',
+        TinyState.withNoneditableRootEditor(hook.editor(), () => {
+          testApplyInlineListFormat({
+            format: 'bold',
+            rawInput: '<p>a</p><ul><li>b</li><li>c</li></ul><p>d</p>',
+            selection: { startPath: [ 0, 0 ], soffset: 0, finishPath: [ 2, 0 ], foffset: 1 },
+            expected: '<p>a</p><ul><li>b</li><li>c</li></ul><p>d</p>',
+          });
         });
-        editor.getBody().contentEditable = 'true';
       });
     });
   });
@@ -352,17 +350,16 @@ describe('browser.tinymce.core.fmt.ListItemFormatTest', () => {
       );
 
       it('TINY-9563: remove bold on LI elements in a noneditable root should not remove bold styles', () => {
-        const editor = hook.editor();
-        const initialContent = '<ul><li style="font-weight: bold;">b</li><li style="font-weight: bold;">c</li></ul>';
+        TinyState.withNoneditableRootEditor(hook.editor(), () => {
+          const initialContent = '<ul><li style="font-weight: bold;">b</li><li style="font-weight: bold;">c</li></ul>';
 
-        editor.getBody().contentEditable = 'false';
-        testRemoveInlineListFormat({
-          format: 'bold',
-          rawInput: initialContent,
-          selection: { startPath: [ 0, 0, 0 ], soffset: 0, finishPath: [ 0, 1, 0 ], foffset: 1 },
-          expected: initialContent,
+          testRemoveInlineListFormat({
+            format: 'bold',
+            rawInput: initialContent,
+            selection: { startPath: [ 0, 0, 0 ], soffset: 0, finishPath: [ 0, 1, 0 ], foffset: 1 },
+            expected: initialContent,
+          });
         });
-        editor.getBody().contentEditable = 'true';
       });
     });