+++ modules/tinymce/src/plugins/emoticons/test/ts/browser/NoneditableRootTest.ts
@@ -1,0 +1,40 @@
+import { Keys, UiFinder } from '@ephox/agar';
+import { describe, it } from '@ephox/bedrock-client';
+import { SugarBody } from '@ephox/sugar';
+import { TinyHooks, TinySelections, TinyState, TinyUiActions } from '@ephox/wrap-mcagar';
+
+import Editor from 'tinymce/core/api/Editor';
+import Plugin from 'tinymce/plugins/emoticons/Plugin';
+
+describe('browser.tinymce.plugins.emoticons.NoneditableRootTest', () => {
+  const hook = TinyHooks.bddSetup<Editor>({
+    plugins: 'emoticons',
+    toolbar: 'emoticons',
+    base_url: '/project/tinymce/js/tinymce'
+  }, [ Plugin ], true);
+
+  it('TINY-9669: Disable emoticons button on noneditable content', () => {
+    TinyState.withNoneditableRootEditor(hook.editor(), (editor) => {
+      editor.setContent('<div>Noneditable content</div><div contenteditable="true">Editable content</div>');
+      TinySelections.setSelection(editor, [ 0, 0 ], 0, [ 0, 0 ], 2);
+      UiFinder.exists(SugarBody.body(), '[aria-label="Emojis"][aria-disabled="true"]');
+      TinySelections.setSelection(editor, [ 1, 0 ], 0, [ 1, 0 ], 2);
+      UiFinder.exists(SugarBody.body(), '[aria-label="Emojis"][aria-disabled="false"]');
+    });
+  });
+
+  it('TINY-9669: Disable emoticons menuitem on noneditable content', async () => {
+    await TinyState.withNoneditableRootEditorAsync(hook.editor(), async (editor) => {
+      editor.setContent('<div>Noneditable content</div><div contenteditable="true">Editable content</div>');
+      TinySelections.setSelection(editor, [ 0, 0 ], 0, [ 0, 0 ], 2);
+      TinyUiActions.clickOnMenu(editor, 'button:contains("Insert")');
+      await TinyUiActions.pWaitForUi(editor, '[role="menuitem"][title="Emojis..."][aria-disabled="true"]');
+      TinyUiActions.keystroke(editor, Keys.escape());
+      TinySelections.setSelection(editor, [ 1, 0 ], 0, [ 1, 0 ], 2);
+      TinyUiActions.clickOnMenu(editor, 'button:contains("Insert")');
+      await TinyUiActions.pWaitForUi(editor, '[role="menuitem"][title="Emojis..."][aria-disabled="false"]');
+      TinyUiActions.keystroke(editor, Keys.escape());
+    });
+  });
+});
+