+++ modules/tinymce/src/themes/silver/main/ts/ui/core/IndentOutdent.ts
@@ -1,11 +1,11 @@
 import Editor from 'tinymce/core/api/Editor';
 import { Toolbar } from 'tinymce/core/api/ui/Ui';
 
-import { onActionExecCommand, onSetupEvent } from './ControlUtils';
+import { onActionExecCommand, onSetupEditableToggle, onSetupEvent } from './ControlUtils';
 
 const onSetupOutdentState = (editor: Editor) =>
   onSetupEvent(editor, 'NodeChange', (api: Toolbar.ToolbarButtonInstanceApi) => {
-    api.setEnabled(editor.queryCommandState('outdent'));
+    api.setEnabled(editor.queryCommandState('outdent') && editor.selection.isEditable());
   });
 
 const registerButtons = (editor: Editor): void => {
@@ -19,6 +19,7 @@ const registerButtons = (editor: Editor): void => {
   editor.ui.registry.addButton('indent', {
     tooltip: 'Increase indent',
     icon: 'indent',
+    onSetup: onSetupEditableToggle(editor),
     onAction: onActionExecCommand(editor, 'indent')
   });
 };